Changes in dafault statisticsComputer
CHANGE: Now statistics are computed from stream of observables instead
of stream of values
ADD: Statistics are now compatible with non numeric types (string). If
value is a string and string value couldn't be converted to a number, it
will handle string length as its value. Ex.: "AB" will handle as 2